refactor: split flow into steps
