+++ packages/shipjs/src/color.js
@@ -1,9 +1,8 @@
 import chalk from 'chalk';
 
-export const info = chalk.keyword('cyan');
-export const warning = chalk.keyword('yellow');
-export const error = chalk.keyword('red');
-export const magenta = chalk.keyword('magenta');
-export const grey = chalk.grey;
+export const info = chalk.keyword('cyan').dim;
+export const warning = chalk.keyword('darkorange');
+export const error = chalk.keyword('tomato');
+export const slateblue = chalk.keyword('slateblue');
 export const bold = chalk.bold;
 export const underline = chalk.underline;
