+++ packages/shipjs/src/step/finished.js
@@ -1,0 +1,3 @@
+import runStep from './runStep';
+
+export default () => runStep({ title: 'All Finished.' }, () => {});
