+++ packages/shipjs/src/step/prepare/checkoutToStagingBranch.js
@@ -1,0 +1,6 @@
+import runStep from '../runStep';
+
+export default ({ stagingBranch, dir, dryRun }) =>
+  runStep({ title: 'Checking out to the staging branch.' }, ({ run }) => {
+    run(`git checkout -b ${stagingBranch}`, dir, dryRun);
+  });
