+++ packages/shipjs/src/step/pull.js
@@ -1,0 +1,6 @@
+import runStep from './runStep';
+
+export default ({ dir, dryRun }) =>
+  runStep({ title: 'Updating from remote.' }, ({ run }) => {
+    run('git pull', dir, dryRun);
+  });
