+++ packages/shipjs/src/step/push.js
@@ -1,0 +1,6 @@
+import runStep from './runStep';
+
+export default ({ dir, dryRun }) =>
+  runStep({ title: 'Pushing to remote.' }, ({ run }) => {
+    run('git pull', dir, dryRun);
+  });
