+++ packages/shipjs/src/step/release/runBuild.js
@@ -1,0 +1,7 @@
+import runStep from '../runStep';
+
+export default ({ isYarn, config, dir, dryRun }) =>
+  runStep({ title: 'Building.' }, ({ run }) => {
+    const { buildCommand } = config;
+    run(buildCommand({ isYarn }), dir, dryRun);
+  });
