+++ packages/shipjs/src/step/release/runTest.js
@@ -1,0 +1,7 @@
+import runStep from '../runStep';
+
+export default ({ isYarn, config, dir, dryRun }) =>
+  runStep({ title: 'Running test.' }, ({ run }) => {
+    const { testCommandBeforeRelease } = config;
+    run(testCommandBeforeRelease({ isYarn }), dir, dryRun);
+  });
