--- packages/shipjs/src/util/printStep.js
@@ -1,6 +1,0 @@
-import { magenta, bold } from '../color';
-import print from './print';
-
-export default function printStep(message) {
-  print(magenta(bold(`- ${message}`)));
-}
